+++ python/libcasm/container/__init__.py
@@ -1,4 +1,0 @@
-"""libcasm.container: CASM containers, part of libcasm-global"""
-from ._container import *
-
-something = True